[#14] rename : OAuthProvider 폴더 위치 변경
