+++ src/main/java/com/jinddung2/givemeticon/oauth/domain/oauth/OAuthProvider.java
@@ -1,0 +1,5 @@
+package com.jinddung2.givemeticon.oauth.domain.oauth;
+
+public enum OAuthProvider {
+    KAKAO, NAVER;
+}
